()Cleaned up utils, added a subutils folder. All CRUD features work server side
