+++ app/server/data/blog_posts/Adam/do_slugs_keep_the_pool_clean
@@ -1,0 +1,1 @@
+slugs are awesome? Maybe they are and maybe they ain't. I don't know anymore
