+++ app/server/route/route_config.js
@@ -42,10 +42,6 @@ exports.utils = {
     Utils.fetch_posts(req,res);
   },
 
-  edit_post: function(req, res) {
-    Utils.edit_post(req, res)
-  },
-
   delete_post: function(req, res) {
     Utils.delete_post(req, res);
   }
